requirements.txt
pandas
matplotlib
Pillow
seaborn
numpy
cmake
opencv-python
scikit-learn
joblib
torch
torchvision
ipywidgets